[FIX] mail: show ellipsis icon in message email read more/less
Before this commit, when a message is of type "email" and contains
parts are in foldable with more/less (like `<quote>`), the icon of
the button for fold/unfold is missing.

This happens because the icon is `.oi` and requires `odoo_ui_icons`.
Message of type "email" have their content inside a shadow DOM,
because we want to preserve the style of email inside the web client,
at least in white theme. Shadow-DOM prevents the parent document to
pass `odoo_ui_icons` thus the Shadow-DOM could not apply expected
style on `.oi` icons.

This commit fixes the issue by passing the required CSS as stylesheet
to the shadow DOM, so that `.oi` icons are working inside the shadow
DOM of message content of type "email".

closes odoo/odoo#236421

X-original-commit: b3878d0ce9204f20ff5e8822c4d140207cdb1342
