+++ addons/mail/__manifest__.py
@@ -210,6 +210,9 @@ For more specific needs, you may also assign custom-defined actions
         'mail.assets_lamejs': [
             'mail/static/lib/lame/lame.js',
         ],
+        "mail.assets_message_email": [
+            "web/static/lib/odoo_ui_icons/style.css",
+        ],
         'mail.assets_public': [
             'web/static/lib/jquery/jquery.js',
             'web/static/lib/odoo_ui_icons/style.css',
